Refactor packages and add different content classes
