+++ app/src/main/java/inge5a/polytech/com/controller/GameController.java
@@ -1,5 +1,6 @@
 package inge5a.polytech.com.controller;
 
+
 /**
  * Created by thibault on 10/11/17.
  */
@@ -9,8 +10,8 @@ public class GameController {
 
 
     public void play(){
-        while team.score>=0{
+      //  while (team.score>=0) {
 
-        }
+      //  }
     }
 }
