+++ app/src/main/java/inge5a/polytech/com/model/content/ContentImage.java
@@ -1,0 +1,8 @@
+package inge5a.polytech.com.model.content;
+
+/**
+ * Created by nico- on 18/10/2017.
+ */
+
+public class ContentImage extends Content {
+}
