+++ app/src/main/java/inge5a/polytech/com/model/question/QuestionText.java
@@ -1,6 +1,9 @@
-package inge5a.polytech.com.model;
+package inge5a.polytech.com.model.question;
 
 import java.util.ArrayList;
+
+import inge5a.polytech.com.model.Answer;
+import inge5a.polytech.com.model.content.Content;
 
 /**
  * Created by thibault on 10/11/17.
